Add header file for GNUmakeTokenPoolPosix
This allows actually using this implementation.
